Various changes - see desc
Changed the sound box to use packed scenes that utilize the new OneShotAudio class

Added the sound box which is a class that holds various SFX that can be triggered randomly within the collection or by index.

Added a combined DialogSoundBox Prefab - combines both an audioBox and passive dialog. When an index is called it will trigger the dialog and sound at that index. If the index doesn't exist for one or the other then it will be skipped and the other will play.